requirements.txt
streamlit
pandas
numpy
matplotlib
seaborn
plotly
openpyxl  # For Excel support
psycopg2-binary  # PostgreSQL connector